✨feat: deleteComment 구현

+++ src/main/java/com/project/likelion13thbe/domain/comment/service/command/CommentCommandServiceImpl.java
@@ -65,9 +65,14 @@ public class CommentCommandServiceImpl implements CommentCommandService {
     }
 
     @Override
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, String email) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new CommentException(CommentErrorCode.COMMENT_NOT_FOUND));
+
+        if (!comment.getMember().getEmail().equals(email)) {
+            throw new CommentException(CommentErrorCode.COMMENT_ACCESS_DENIED);
+        }
+
         comment.delete();
     }
 }
